Add files via upload
Assignment 5 is a program coded in C which allows a user to input nodes into a binary search tree. The program will then print the in-order and pre-order traversals of the binary search tree. The user may also search the tree for a specific node. If the node is found, then the program will display the data placed inside the node to the user.